backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
passlib==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
pymysql==1.1.1
python-multipart==0.0.9
